backend/requirements.txt
fastapi
uvicorn
supabase
sentence-transformers
torch
pillow
python-multipart
python-dotenv
requests
uvicorn
